Helper add to remove regex results from string.
Function add to remove the match results of a regular expression in the
base string. This is done in situations where you want to use both a
preg_match_all and preg_replace. Downside of using preg_replace is that
results are not returned as array and regular expression overhead.