Implemented Checkboxes Extra Feature
This feature, when activated through the paperwork configuration
setting, allows a quick way to create disabled checkboxes in the
note view by starting a line with [] or [X] depending on the state.
These checkboxes are not clickable and in the edit view are
transformed again in [] or [X]. A tooltip explaining the feature
is added to the checkboxes.

This could be a plugin but since there's no implementation for them,
this could be implemented in core.