Make the compose command follow the image default
`docker-compose.yml` will default to using the command in the image if
one is not supplied. An issue arose when the location of the command to
run was changed, but only in `Dockerfile`. A clean solution would be to
let the docker-compose use the image default, rather than change it both
places every time.

Also, a white space was removed from the `Dockerfile` as docker build
was throwing warnings about empty lines following `\\` and stating it
would become an error soon:

```
[WARNING]: Empty continuation lines will become errors in a future
release.
```